# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Roguelike API",
        "description": "Api documentation for roguelike game.",
        "termsOfService": "",
        "version": "1.0.0"
    },
    "host": "rogueapi.keisenb.io",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "lists all friends for current user",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with friends",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Friend"
                            }
                        }
                    },
                    "404": {
                        "description": "no friends in table, or an error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Adds a friend by username.",
                "parameters": [
                    {
                        "name": "friend",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddFriend"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Added friend.",
                        "schema": {
                            "$ref": "#/definitions/Friend"
                        }
                    },
                    "401": {
                        "description": "unauthorized, invalid token, missing token, expired token.",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already friends with user.",
                        "schema": {
                            "properties": {
                                "parameter_name": {
                                    "type": "string",
                                    "default": "error message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unable to find friend by username",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "error message"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Deletes a friend by username.",
                "parameters": [
                    {
                        "name": "friend",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddFriend"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successfully deleted friend."
                    },
                    "401": {
                        "description": "unauthorized, invalid token, missing token, expired token.",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unable to find friend by username.",
                        "schema": {
                            "properties": {
                                "parameter_name": {
                                    "type": "string",
                                    "default": "error message"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not friends with user with the display name.",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "error message"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AddFriend": {
            "type": "object",
            "allOf": [
                {
                    "properties": {
                        "display_name": {
                            "type": "string"
                        }
                    }
                }
            ]
        },
...
        "Friend": {
            "type": "object",
            "allOf": [
                {
                    "properties": {
                        "id1": {
                            "type": "integer"
                        },
                        "id2": {
                            "type": "integer"
                        },
                        "user1": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "user2": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            ]
        },
...
    }
}